rename: 같은 관심사 파일들 하나의 폴더로 통합
