+++ src/pages/Hi.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import EmojiBadge from '@components/common/EmojiBadge';
+import EmojiBadge from '@components/common/badge/EmojiBadge';
 
 function Hi() {
   return (
